Migrate write_mod/2 to Cachex.Actions
